+++ src/main/java/it/vitalegi/archi/App.java
@@ -1,7 +1,7 @@
 package it.vitalegi.archi;
 
 import it.vitalegi.archi.diagram.DiagramFormat;
-import it.vitalegi.archi.exporter.plantuml.PlantumlDiagramExporter;
+import it.vitalegi.archi.exporter.c4.plantuml.PlantumlDiagramExporter;
 import it.vitalegi.archi.model.builder.WorkspaceDirector;
 import it.vitalegi.archi.workspaceloader.FileSystemWorkspaceLoader;
 import org.slf4j.Logger;
