+++ src/main/java/it/vitalegi/archi/exporter/c4/plantuml/PlantumlDiagramExporter.java
@@ -1,4 +1,4 @@
-package it.vitalegi.archi.exporter.plantuml;
+package it.vitalegi.archi.exporter.c4.plantuml;
 
 import it.vitalegi.archi.diagram.DiagramFormat;
 import it.vitalegi.archi.model.Workspace;
