+++ src/main/java/it/vitalegi/archi/exporter/c4/plantuml/builder/DiagramElementFactoryVisitor.java
@@ -1,4 +1,4 @@
-package it.vitalegi.archi.exporter.plantuml.builder;
+package it.vitalegi.archi.exporter.c4.plantuml.builder;
 
 import it.vitalegi.archi.model.Model;
 import it.vitalegi.archi.model.diagramelement.C4DiagramElement;
