+++ src/test/java/it/vitalegi/archi/ExampleTests.java
@@ -1,7 +1,7 @@
 package it.vitalegi.archi;
 
 import it.vitalegi.archi.model.builder.WorkspaceDirector;
-import it.vitalegi.archi.exporter.plantuml.PlantumlDiagramExporter;
+import it.vitalegi.archi.exporter.c4.plantuml.PlantumlDiagramExporter;
 import it.vitalegi.archi.diagram.DiagramFormat;
 import it.vitalegi.archi.workspaceloader.FileSystemWorkspaceLoader;
 import org.junit.jupiter.api.Test;
