+++ src/test/java/it/vitalegi/archi/diagram/DiagramValidatorVisitorTests.java
@@ -1,6 +1,6 @@
 package it.vitalegi.archi.diagram;
 
-import it.vitalegi.archi.exporter.plantuml.PlantumlDiagramValidatorVisitor;
+import it.vitalegi.archi.exporter.c4.plantuml.PlantumlDiagramValidatorVisitor;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
